Flag to disable HTTP2 for GCS (#58)
HTTP2 can reuse a connection for multiple requests. If a connection is
made to a bad frontend this can make even retries fail since they use
the same connection which is the root issue. For now gate this behind a
`--disable-http2` flag until we're sure it's stable to be a default.

Tested manually with downloads of GCS files and ensuring `lsof -i -P |
grep fastar` shows multiple connections open.